added copy paste (with bugs of course, see description)
transform of selected objects is copied by reference and not by value, resulting in pasting on the current location of the copied objects. instead of the location at the time of the copy.
also the copy and pasting can not be undone at the moment. we could add a delete function but that would also not be able to be undone.