Saving to JSON format implemented
In SaveLoadManager.cs is the class that will save the current projects data. This class contains multiple custom datatypes for storing unsupported object data in JSON. 

later commits will include loading functionality and saving to a file. currently only saves to a string.
Added  JsonDotNet/documentation and sourcefiles to Gitignore

these files are not required to have the application work