Object Naming, keyboard blocking Fixed , file loading now deletes current scene
Objects now have names following (#) at the end. Like Cube, Cube(1), Cube(2), Cube(2)(1) etc.

Shortcuts that change the selected tool or delete the selected object are now blocked while focused on an InputField.

Save file Loading now deletes all previous objects in the scene to start of with a clean slate. Might want to add a warning when doing so.